Remove whitespace from toggle transformer
In the toggle transformer, there is whitespace before the child content in the block. This is an issue when the child block is a code block, as it causes the first line of the code block to be indented, causing indentation error/inconsistencies in the actual code/text inside the block.