Remove leading whitespace in toggle transformer
In the toggle transformer, there is whitespace before the child content in the block. This is an issue, when the child block is a code block. It causes the first line of the code block to be indented, causing indentation error/inconsistencies in the rendered code/text inside the block.

This PR removes the whitespace.